+++ build_journal_entry.py
@@ -1,7 +1,7 @@
 from datetime import date
 
 from exceptions import JournalEntryValueError
-from parse_journal_entry import RawJournalEntry
+from hparser.parse_journal_entry import RawJournalEntry
 from journal.journal_entry import JournalEntry
 
 def build_journal_entry(raw_journal_entry: RawJournalEntry, record_date: date) -> JournalEntry:
