+++ setup.cfg
@@ -1,0 +1,7 @@
+[metadata]
+name = local_habit_tracker
+version = 0.0.1
+author = szabikr
+
+[options]
+packages = hparser
